refactor: BankTask enum 타입 생성
- LoanOperation과 DepositOperation 타입 병합
- BankOperation에 프로퍼티 추가
- task의 타입을 String에서 BankTask로 변경
