+++ BankManagerConsoleApp/BankManagerConsoleApp/Customer/Customer.swift
@@ -5,5 +5,5 @@
 //  Created by 휴, 보리 on 2022/07/01.
 //
 struct Customer {
-    let task: String
+    let task: BankTask
 }
